+++ RELEASE-NOTES.md
@@ -35,7 +35,8 @@ END TEMPLATE-->
 
 ### Breaking changes
 
-*None yet*
+* Changed default for `net.buffer_size` to `2`.
+* Changed default for `auth.mode` to `Required`. On development builds, the default is overriden to remain at `Optional`, so this only affects published servers.
 
 ### New features
 
@@ -47,7 +48,7 @@ END TEMPLATE-->
 
 ### Other
 
-*None yet*
+* Cleaned up some `FULL_RELEASE` CVar default value overrides into `CVarDefaultOverrides.cs`.
 
 ### Internal